+++ src/Reader.ts
@@ -418,6 +418,6 @@ export const traverseArray: <R, A, B>(
  *
  * @since 3.0.0
  */
-export const sequenceArray: <R, A>(arr: ReadonlyArray<Reader<R, A>>) => Reader<R, ReadonlyArray<A>> = traverseArray(
-  identity
-)
+export const sequenceArray: <R, A>(arr: ReadonlyArray<Reader<R, A>>) => Reader<R, ReadonlyArray<A>> =
+  /*#__PURE__*/
+  traverseArray(identity)