+++ src/TaskEither.ts
@@ -1046,6 +1046,6 @@ export const traverseSeqArray: <A, B, E>(
  *
  * @since 3.0.0
  */
-export const sequenceSeqArray: <A, E>(
-  arr: ReadonlyArray<TaskEither<E, A>>
-) => TaskEither<E, ReadonlyArray<A>> = traverseSeqArray(identity)
+export const sequenceSeqArray: <A, E>(arr: ReadonlyArray<TaskEither<E, A>>) => TaskEither<E, ReadonlyArray<A>> =
+  /*#__PURE__*/
+  traverseSeqArray(identity)